# backend/models/swagger.go
package models

// SwaggerResponse represents a generic response object
type SwaggerResponse struct {
	Message string `json:"message,omitempty"`
	Error   string `json:"error,omitempty"`
}

// ErrorResponse represents an error response
type ErrorResponse struct {
	Error string `json:"error" example:"error message"`
}